requirements.txt
Django
drf-yasg
djangorestframework~=3.15.2
django-filter~=24.3
gunicorn==23.0.0
dj-database_url
psycopg2-binary
django-environ
django-cors-headers
pillow==11.0.0